backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.26.0
google-genai>=0.8.0
pydantic>=2.7.0
python-multipart>=0.0.9
redis>=5.0.0
supabase>=2.5.1
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
pytest>=8.2.0
